components/backend-api/requirements.txt
google-cloud-firestore
google-cloud-logging
